Validaciones

+++ web/Controllers/CustomerController.cs
@@ -27,8 +27,6 @@ namespace web.Controllers
                 ViewBag.customers = clientes;
                 return View();
             }
-            return View();
-
         }
         [HttpPost]
         public IActionResult SaveCustomer(Customer customer)
@@ -63,36 +61,39 @@ namespace web.Controllers
             return NotFound();
         }
         [HttpPost]
-        public JsonResult DeleteCustomer(int id)
-        {
-            bool resultado = false;
-            var customer = _context.Customers.FirstOrDefault(s => s.Idcustomer == id);
-            if (customer != null)
-            {
-                _context.Customers.Remove(customer);
-                _context.SaveChanges();
-                resultado = true;
-            }
-            return Json(resultado);
-        }
+        [ValidateAntiForgeryToken]
         public IActionResult UpdateCustomer(Customer customer)
         {
-           
-            if (ModelState.IsValid)
+            if(ModelState.IsValid)
             {
-                var update = new Customer()
-                {
-                    Idcustomer = customer.Idcustomer,
-                    Namecustomer = customer.Namecustomer,
-                    Lastnamecustomer = customer.Lastnamecustomer,
-                    Phonecustomer = customer.Phonecustomer,
-                    Dnicustomer = customer.Dnicustomer
-                };
-                _context.Update(update);
+                _context.Customers.Update(customer);
                 _context.SaveChanges();
-                return RedirectToAction("Index");
             }
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction("Index");
+        }
+        [HttpPost]
+        public JsonResult Confirmar(int id)
+        {
+            bool estado = true;
+            List<Sale> ver = _context.Sales.Where(x => x.Idcustomer == id).ToList();
+            if (ver.Count > 0)
+            {
+                estado = false;
+                return Json(estado);
+            }
+            else
+            {
+                var customer = _context.Customers.FirstOrDefault(s => s.Idcustomer == id);
+                if (customer != null)
+                {
+                    _context.Customers.Remove(customer);
+                    _context.SaveChanges();
+                    estado = true;
+                }
+                
+                return Json(estado);
+            }
+            
         }
     }
 }
